refactor(convert): extract table module with try_convert entry
Phase 1 of fulgur-odgo. Move 3 table helpers from convert/mod.rs into
convert/table.rs (all private): convert_table, is_table_section,
collect_table_cells.

Add pub(super) fn try_convert(doc, node_id, ctx, depth) -> Option<...>
wrapping the dispatcher's <table> branch. Signature includes depth
(deviates from the plan template since convert_table needs it for
child recursion).

The dispatcher's elem_data block — which had only the table check left —
collapses entirely: table::try_convert now subsumes the element-data
match. A breadcrumb comment in mod.rs notes table::try_convert covers
<table>, mirroring the replaced::try_convert breadcrumb.

No behavioral change. VRT byte-identical against committed goldens.
